Improvements to cookie banner tests:
Smoke tests - run against news & pashto services
Non smoke tests - run against all configured page types where smoke is true ~ 25 URLs
